Fixed gyroscope to read high and low registers
The data coming out of this input seems to still have an accuracy problem but it is decently precise. 
You can probably use the gyro in this state to make a pretty good gyroscope library/function set but we need to figure out where that slight inaccuracy is coming from. 
The output seems to have a spike every now and then but I can't find the cause just yet.