AI Camera Autodetect
This update allows the camera code in libwallaby to autodetect the
AI camera.  It does this by attempting to send a status request to
the AI Camera command server. If there is any failure, it assumes
that the AI camera is either not there or not ready.

This patch requires the patch in pull request #4 for ai-rpi-cam to
be installed in the AI camera software.